Organise route import into single file
